Original change: https://android-review.googlesource.com/c/platform/packages/modules/NeuralNetworks/+/1778651

+++ common/CpuExecutor.cpp
@@ -269,7 +269,7 @@ bool OperationExecutionContext::checkNoZeroSizedInput() const {
     for (uint32_t i = 0; i < operation->inputs.size(); i++) {
         if (isOmittedInput(i)) continue;
         for (uint32_t j = 0; j < getInputInfo(i)->dimensions.size(); j++) {
-            NN_RET_CHECK_NE(getInputInfo(i)->dimensions[j], 0)
+            NN_RET_CHECK_NE(getInputInfo(i)->dimensions[j], 0u)
                     << operation->type << " does not support zero-sized tensor, but input " << i
                     << " dimension " << j << " is 0.";
         }
@@ -1375,7 +1375,7 @@ int CpuExecutor::executeOperation([[maybe_unused]] const Operation& operation,
             const int32_t axis = getScalarData<int32_t>(operands[ins[1]]);
             const int32_t numOutputs = getScalarData<int32_t>(operands[ins[2]]);
 
-            if (numOutputs != outs.size()) {
+            if (static_cast<size_t>(numOutputs) != outs.size()) {
                 return ANEURALNETWORKS_BAD_DATA;
             }
